requirements.txt
streamlit-bridge